refactor: move trending podcast section and filter out of feature components
