+++ frontend/src/pages/podcast/PodcastCategoryPage/PodcastCategoryPage.tsx
@@ -1,7 +1,7 @@
 import "./PodcastCategoryPage.css"
 import { useCallback, useEffect, useMemo, useState } from "react"
 import { useNavigate, useParams } from "react-router"
-import TrendingPodcastSection from "../../../features/podcast/trending/components/TrendingPodcastSection/TrendingPodcastSection.tsx"
+import TrendingPodcastSection from "../../../features/podcast/trending/TrendingPodcastSection/TrendingPodcastSection.tsx"
 import useTrendingPodcasts from "../../../hooks/podcast/useTrendingPodcasts.ts"
 import Breadcrumb from "../../../components/ui/breadcrumb/index.tsx"
